modules/py_image: Move statistics objects to py_image_stats.c.
Moves the similarity, statistics, percentile, threshold and histogram
objects into a separate file to reduce the size of py_image.c.
